social_media_api/pyproject.toml
[tool.poetry]
name = "social-media-api"
version = "0.1.0"
description = ""
authors = ["j1ere <jeremywambua106@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
